+++ docs/cookbook/components/script.composition-api.ts
@@ -1,4 +1,4 @@
-import { createComponent, computed, reactive, PropType } from '@vue/composition-api'
+import { createComponent, computed, ref } from '@vue/composition-api'
 
 interface User {
   firstName: string
@@ -8,14 +8,14 @@ interface User {
 export default createComponent({
   props: {
     user: {
-      type: Object as PropType<User>,
+      type: Object as () => User,
       required: true
     }
   },
 
   setup ({ user }) {
     const fullName = computed(() => `${user.firstName} ${user.lastName}`)
-    const message = reactive('This is a message')
+    const message = ref('This is a message')
 
     return {
       fullName,