+++ jest.config.js
@@ -1,9 +1,19 @@
 module.exports = {
   testEnvironment: 'node',
+  transform: {
+    '^.+\\.ts$': 'ts-jest'
+  },
+  globals: {
+    'ts-jest': {
+      tsConfig: 'tsconfig.test.json',
+      diagnostics: {
+        ignoreCodes: [2345]
+      }
+    }
+  },
   collectCoverage: true,
-  coverageDirectory: './coverage',
   collectCoverageFrom: [
-    'packages/*/lib/**/*.js'
+    'packages/*/src/**/*.ts'
   ],
   coverageThreshold: {
     global: {
@@ -12,14 +22,5 @@ module.exports = {
       lines: 100,
       statements: 100
     }
-  },
-  moduleNameMapper: {
-    '^~/(.*)$': '<rootDir>/lib/$1',
-    '^~~$': '<rootDir>',
-    '^@@$': '<rootDir>',
-    '^@/(.*)$': '<rootDir>/lib/$1'
-  },
-  transform: {
-    '^.+\\.js$': 'babel-jest'
   }
 }