+++ packages/types/app/index.d.ts
@@ -13,6 +13,7 @@ type NuxtState = Record<string, any>
 
 export interface Context {
   app: NuxtAppOptions
+  base: string
   /**
    * @deprecated Use process.client instead
   */
@@ -28,6 +29,7 @@ export interface Context {
   isDev: boolean
   isHMR: boolean
   route: Route
+  from: Route
   store: Store<any>
   env: Record<string, any>
   params: Route['params']
@@ -98,4 +100,12 @@ export interface NuxtApp extends Vue {
   loadLayout(layout: string): Promise<any> // TBD
   refresh(): void
   setLayout(layout: string): any // TBD
+
+}
+
+// window.$nuxt
+declare global {
+  interface Window {
+    $nuxt: NuxtApp
+  }
 }