+++ packages/types/config/build.d.ts
@@ -3,23 +3,86 @@
  * Documentation: https://nuxtjs.org/api/configuration-build
  */
 
+import { TransformOptions, PluginItem } from '@babel/core'
+import { Options as AutoprefixerOptions } from 'autoprefixer'
+import { Options as FileLoaderOptions } from 'file-loader'
+import { Options as HtmlMinifierOptions } from 'html-minifier'
+import * as Less from 'less'
+import { Options as SassOptions } from 'node-sass'
+import { Options as OptimizeCssAssetsWebpackPluginOptions } from 'optimize-css-assets-webpack-plugin'
+import { Plugin as PostcssPlugin } from 'postcss'
+import { Options as PugOptions } from 'pug'
+import { TerserPluginOptions } from 'terser-webpack-plugin'
+import { VueLoaderOptions } from 'vue-loader'
 import {
   Configuration as WebpackConfiguration,
+  Loader as WebpackLoader,
+  loader as WebpackLoaderNamespace,
   Options as WebpackOptions,
   Plugin as WebpackPlugin
 } from 'webpack'
-import { TransformOptions, PluginItem } from '@babel/core'
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
 import { Options as WebpackDevMiddlewareOptions } from 'webpack-dev-middleware'
 import { Options as WebpackHotMiddlewareOptions } from 'webpack-hot-middleware'
-import { Options as HtmlMinifierOptions } from 'html-minifier'
-import { Options as OptimizeCssAssetsWebpackPluginOptions } from 'optimize-css-assets-webpack-plugin'
-import { TerserPluginOptions } from 'terser-webpack-plugin'
 
-type NuxtConfigurationLoaders = any // TBD
+type CssLoaderUrlFunction = (url: string, resourcePath: string) => boolean
+type CssLoaderImportFunction = (parsedImport: string, resourcePath: string) => boolean
+type CssLoaderMode = 'global' | 'local'
+interface CssLoaderModulesOptions {
+  context?: string
+  getLocalIdent?: (context: string, localIdentName: string, localName: string, options: CssLoaderModulesOptions) => string
+  hashPrefix?: string
+  localIdentName?: string
+  localIdentRegExp?: string | RegExp
+  mode?: CssLoaderMode
+}
+
+interface CssLoaderOptions {
+  import?: boolean | CssLoaderImportFunction
+  importLoaders?: number
+  localsConvention?: 'asIs' | 'camelCase' | 'camelCaseOnly' | 'dashes' | 'dashesOnly'
+  modules?: boolean | CssLoaderMode | CssLoaderModulesOptions
+  onlyLocals?: boolean
+  sourceMap?: boolean
+  url?: boolean | CssLoaderUrlFunction
+}
+
+interface UrlLoaderOptions {
+  esModules?: boolean
+  fallback?: WebpackLoader
+  limit?: boolean | number | string
+  mimetype?: string
+}
+
+interface NuxtConfigurationLoaders {
+  css?: CssLoaderOptions
+  cssModules?: CssLoaderOptions
+  file?: FileLoaderOptions
+  fontUrl?: UrlLoaderOptions
+  imgUrl?: UrlLoaderOptions
+  less?: Less.Options
+  pugPlain?: PugOptions
+  sass?: SassOptions
+  scss?: SassOptions
+  stylus?: any // TBD
+  vue?: VueLoaderOptions
+  vueStyle?: {
+    manualInject?: boolean
+    ssrId?: boolean
+    shadowMode?: boolean
+  }
+}
+
+interface NuxtWebpackEnv {
+  isClient: boolean
+  isDev: boolean
+  isLegacy: boolean
+  isModern: boolean
+  isServer: boolean
+}
 
 interface NuxtBabelPresetEnv {
-  isServer: boolean
+  envName: 'client' | 'modern' | 'server'
 }
 
 interface NuxtBabelOptions extends Pick<TransformOptions, Exclude<keyof TransformOptions, 'presets'>> {
@@ -27,10 +90,49 @@ interface NuxtBabelOptions extends Pick<TransformOptions, Exclude<keyof Transfor
   cacheDirectory?: boolean
   cacheIdentifier?: string
   customize?: string | null
-  presets?: ((env: NuxtBabelPresetEnv, defaultPreset: [string, object]) => PluginItem[] | void) | PluginItem[] | null
+  presets?: ((env: NuxtBabelPresetEnv & NuxtWebpackEnv, defaultPreset: [string, object]) => PluginItem[] | void) | PluginItem[] | null
+}
+
+interface Warning {
+  message: string
+  name: string
+}
+
+interface PostcssOrderPresetFunctions {
+  cssnanoLast: (names: string[]) => string[]
+  presetEnvAndCssnanoLast: (names: string[]) => string[]
+  presetEnvLast: (names: string[]) => string[]
+}
+type PostcssOrderPreset = keyof PostcssOrderPresetFunctions
+interface PostcssVariableMap {
+  customMedia: Record<string, string>
+  customProperties: Record<string, string>
+  customSelectors: Record<string, string>
+  environmentVariables?: Record<string, string>
+}
+
+interface PostcssConfiguration {
+  order?: PostcssOrderPreset | string[] | ((names: string[], presets: PostcssOrderPresetFunctions) => string[])
+  plugins?: {
+    [key: string]: false | { [key: string]: any }
+  } | ((loader: WebpackLoaderNamespace.LoaderContext) => PostcssPlugin<any>[]) | Array<[string | PostcssPlugin<any>, any] | string | PostcssPlugin<any>>
+  preset?: {
+    autoprefixer?: false | AutoprefixerOptions
+    browsers?: string
+    exportTo?: string | string[] | Partial<PostcssVariableMap> | ((map: PostcssVariableMap) => Partial<PostcssVariableMap>)
+    features?: {
+      [key: string]: boolean | { [key: string]: any }
+    }
+    importFrom?: string | string[] | Partial<PostcssVariableMap> | (() => Partial<PostcssVariableMap>)
+    insertAfter?: { [key: string]: PostcssPlugin<any> }
+    insertBefore?: { [key: string]: PostcssPlugin<any> }
+    preserve?: boolean
+    stage?: 0 | 1 | 2 | 3 | 4 | false
+  }
 }
 
 export interface NuxtConfigurationBuild {
+  additionalExtensions?: string[]
   analyze?: BundleAnalyzerPlugin.Options | boolean
   babel?: NuxtBabelOptions
   cache?: boolean
@@ -41,17 +143,14 @@ export interface NuxtConfigurationBuild {
   extend?(
     config: WebpackConfiguration,
     ctx: {
-      isDev: boolean,
-      isClient: boolean,
-      isServer: boolean,
       loaders: NuxtConfigurationLoaders
-    }
+    } & NuxtWebpackEnv
   ): void
   extractCSS?: boolean | Record<string, any>
-  filenames?: { [key in 'app' | 'chunk' | 'css' | 'img' | 'font' | 'video']?: (ctx: { isDev: boolean }) => string }
+  filenames?: { [key in 'app' | 'chunk' | 'css' | 'img' | 'font' | 'video']?: (ctx: NuxtWebpackEnv) => string }
   friendlyErrors?: boolean
   hardSource?: boolean
-  hotMiddleware?: WebpackHotMiddlewareOptions
+  hotMiddleware?: WebpackHotMiddlewareOptions & { client: any /* TBD */ }
   html?: { minify: HtmlMinifierOptions }
   indicator?: boolean
   loaders?: NuxtConfigurationLoaders
@@ -59,7 +158,7 @@ export interface NuxtConfigurationBuild {
   optimizeCSS?: OptimizeCssAssetsWebpackPluginOptions | boolean
   parallel?: boolean
   plugins?: WebpackPlugin[]
-  postcss?: any // TBD
+  postcss?: string[] | boolean | PostcssConfiguration | (() => PostcssConfiguration)
   profile?: boolean
   publicPath?: string
   quiet?: boolean
@@ -72,6 +171,7 @@ export interface NuxtConfigurationBuild {
   standalone?: boolean
   templates?: any
   terser?: TerserPluginOptions | boolean
-  transpile?: (string | RegExp)[]
+  transpile?: (string | RegExp | ((context: NuxtWebpackEnv) => string | RegExp))[]
+  warningIgnoreFilters?: Array<(warn: Warning) => boolean>
   watch?: string[]
 }