+++ packages/types/config/module.d.ts
@@ -4,10 +4,24 @@
  *                https://nuxtjs.org/guide/modules
  */
 
-import { Configuration } from '.'
+import { Configuration as WebpackConfiguration } from 'webpack'
+import { NuxtConfigurationLoaders } from './build'
+import { Configuration as NuxtConfiguration } from '.'
+
+interface ExtendFunctionContext {
+  isClient: boolean
+  isDev: boolean
+  isLegacy: boolean
+  isModern: boolean
+  isServer: boolean
+  loaders: NuxtConfigurationLoaders
+}
+
+type ExtendFunction = (config: WebpackConfiguration, ctx: ExtendFunctionContext) => void
 
 interface ModuleThis {
-  options: Configuration
+  extendBuild(fn: ExtendFunction): void
+  options: NuxtConfiguration
   nuxt: any // TBD
   [key: string]: any // TBD
 }