+++ packages/typescript-build/test/fixture/nuxt.config.js
@@ -1,5 +1,0 @@
-export default {
-  modules: [
-    '../../../packages/typescript-build'
-  ]
-}