+++ packages/typescript-build/test/fixture/nuxt.config.ts
@@ -1,0 +1,13 @@
+import { Configuration } from '@nuxt/types'
+import logger from './serverMiddlewares/logger'
+
+const config: Configuration = {
+  buildModules: [
+    '@nuxt/typescript-build'
+  ],
+  serverMiddleware: [
+    logger
+  ]
+}
+
+export default config