+++ packages/typescript-runtime/lib/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  ...require('./resolve'),
-  ...require('./args'),
-  registerTSNode: require('./register')
-}