+++ packages/types/CHANGELOG.md
@@ -3,6 +3,264 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.4](https://github.com/nuxt/typescript/compare/@nuxt/types@0.5.3...@nuxt/types@0.5.4) (2019-12-09)
+
+**Note:** Version bump only for package @nuxt/types
+
+
+
+
+
+## [0.5.3](https://github.com/nuxt/typescript/compare/@nuxt/types@0.5.2...@nuxt/types@0.5.3) (2019-12-02)
+
+
+### Bug Fixes
+
+* **types:** use connect type for server middleware ([#227](https://github.com/nuxt/typescript/issues/227)) ([0d50ef2](https://github.com/nuxt/typescript/commit/0d50ef27201aca9c749331563dbd6ad0bbf1e5c3)), closes [#183](https://github.com/nuxt/typescript/issues/183)
+
+
+
+
+
+## [0.5.2](https://github.com/nuxt/typescript/compare/@nuxt/types@0.5.1...@nuxt/types@0.5.2) (2019-12-02)
+
+
+### Bug Fixes
+
+* **types:** fix PostcssPlugin type error ([e9a7965](https://github.com/nuxt/typescript/commit/e9a79656ac71e2b9893b9eb6100d87e80636b1a4))
+
+
+
+
+
+## [0.5.1](https://github.com/nuxt/typescript/compare/@nuxt/types@0.5.0...@nuxt/types@0.5.1) (2019-12-02)
+
+
+### Bug Fixes
+
+* **types:** add functions and arrays for plugins ([#225](https://github.com/nuxt/typescript/issues/225)) ([748b465](https://github.com/nuxt/typescript/commit/748b46526e6fc555ba5298a53c1091c9ca980ec3))
+
+
+
+
+
+# [0.5.0](https://github.com/nuxt/typescript/compare/@nuxt/types@0.4.1...@nuxt/types@0.5.0) (2019-12-01)
+
+
+### Bug Fixes
+
+* **types:** improve consistency with class nuxtEnv ([#222](https://github.com/nuxt/typescript/issues/222)) ([2631980](https://github.com/nuxt/typescript/commit/26319806ff4f5b9c81b34f0859e3352365e4e095))
+
+
+### Features
+
+* **types:** add postcss build types ([#223](https://github.com/nuxt/typescript/issues/223)) ([7268990](https://github.com/nuxt/typescript/commit/7268990a06c6d3c8b20d1dbf312e7a3fbfd05cc4))
+
+
+
+
+
+## [0.4.1](https://github.com/nuxt/typescript/compare/@nuxt/types@0.4.0...@nuxt/types@0.4.1) (2019-12-01)
+
+
+### Bug Fixes
+
+* **types:** add correct url-loader options ([#220](https://github.com/nuxt/typescript/issues/220)) ([01084f1](https://github.com/nuxt/typescript/commit/01084f180570f703b6d5583e45d3de4d7fc14dc3)), closes [#221](https://github.com/nuxt/typescript/issues/221)
+
+
+
+
+
+# [0.4.0](https://github.com/nuxt/typescript/compare/@nuxt/types@0.3.4...@nuxt/types@0.4.0) (2019-11-30)
+
+
+### Bug Fixes
+
+* **types:** add `from` property in context interface ([#209](https://github.com/nuxt/typescript/issues/209)) ([cb98916](https://github.com/nuxt/typescript/commit/cb989165a9e98a231d92095af4fdd0f5015614d0))
+* **types:** add babel preset types for function argument ([#192](https://github.com/nuxt/typescript/issues/192)) ([fff6f92](https://github.com/nuxt/typescript/commit/fff6f92b60b1ee7f9665e16972810c3f03029267)), closes [#6691](https://github.com/nuxt/typescript/issues/6691)
+* **types:** fix build.hotMiddleware.client ([7f5d9f8](https://github.com/nuxt/typescript/commit/7f5d9f8b9be802d53286603120145428a9643f57)), closes [#187](https://github.com/nuxt/typescript/issues/187)
+
+
+### Features
+
+* **types:** add types for NuxtConfigurationLoaders ([#202](https://github.com/nuxt/typescript/issues/202)) ([df586d0](https://github.com/nuxt/typescript/commit/df586d0a17cb843a907da89e2ee715257e13e4ce))
+* **types:** add types for NuxtConfigurationRender ([#203](https://github.com/nuxt/typescript/issues/203)) ([bc394de](https://github.com/nuxt/typescript/commit/bc394de9ebac22b7911a15203d247121d85f3aae))
+* **types:** add window.$nuxt type ([63847c7](https://github.com/nuxt/typescript/commit/63847c741103286e61e014e191648346e4b3b275)), closes [#194](https://github.com/nuxt/typescript/issues/194)
+
+
+
+
+
+## [0.3.4](https://github.com/nuxt/typescript/compare/@nuxt/types@0.3.3...@nuxt/types@0.3.4) (2019-11-05)
+
+
+### Bug Fixes
+
+* **types:** add prefix property to NuxtConfigurationServerMiddleware ([#182](https://github.com/nuxt/typescript/issues/182)) ([3f8c54b](https://github.com/nuxt/typescript/commit/3f8c54b6be6efdfb631c54fb726d042d8e89e7df))
+
+
+
+
+
+## [0.3.3](https://github.com/nuxt/typescript/compare/@nuxt/types@0.3.2...@nuxt/types@0.3.3) (2019-11-03)
+
+
+### Bug Fixes
+
+* **types:** add features configuration types ([#160](https://github.com/nuxt/typescript/issues/160)) ([18b07a8](https://github.com/nuxt/typescript/commit/18b07a821d8ead10888507c0f2fc02788642d793))
+* **types:** add router trailingSlash ([#176](https://github.com/nuxt/typescript/issues/176)) ([d3c8677](https://github.com/nuxt/typescript/commit/d3c867795756667120c2562e36bbb1903f99e6ff))
+
+
+
+
+
+## [0.3.2](https://github.com/nuxt/typescript/compare/@nuxt/types@0.3.1...@nuxt/types@0.3.2) (2019-10-18)
+
+**Note:** Version bump only for package @nuxt/types
+
+
+
+
+
+## [0.3.1](https://github.com/nuxt/typescript/compare/@nuxt/types@0.3.0...@nuxt/types@0.3.1) (2019-10-08)
+
+**Note:** Version bump only for package @nuxt/types
+
+
+
+
+
+# [0.3.0](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.15...@nuxt/types@0.3.0) (2019-10-02)
+
+
+### Bug Fixes
+
+* **types:** add generate.exclude type ([9c11abd](https://github.com/nuxt/typescript/commit/9c11abd)), closes [#132](https://github.com/nuxt/typescript/issues/132)
+* **types:** add typing for Routes meta ([#89](https://github.com/nuxt/typescript/issues/89)) ([19df2f9](https://github.com/nuxt/typescript/commit/19df2f9))
+* **types:** remote erroneous color ([#137](https://github.com/nuxt/typescript/issues/137)) ([011c3c1](https://github.com/nuxt/typescript/commit/011c3c1))
+
+
+### Features
+
+* add custom blue color for nuxt cli ([#133](https://github.com/nuxt/typescript/issues/133)) ([3534073](https://github.com/nuxt/typescript/commit/3534073)), closes [#6397](https://github.com/nuxt/typescript/issues/6397)
+
+
+
+
+
+## [0.2.15](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.14...@nuxt/types@0.2.15) (2019-09-23)
+
+
+### Bug Fixes
+
+* **types:** add VueRouter base in Context ([#123](https://github.com/nuxt/typescript/issues/123)) ([29b5702](https://github.com/nuxt/typescript/commit/29b5702))
+
+
+
+
+
+## [0.2.14](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.13...@nuxt/types@0.2.14) (2019-09-17)
+
+
+### Bug Fixes
+
+* **types:** add isModern type definition to build.filenames context ([#120](https://github.com/nuxt/typescript/issues/120)) ([1333259](https://github.com/nuxt/typescript/commit/1333259))
+
+
+
+
+
+## [0.2.13](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.12...@nuxt/types@0.2.13) (2019-09-16)
+
+
+### Bug Fixes
+
+* **types:** allow no return in asyncData ([#116](https://github.com/nuxt/typescript/issues/116)) ([48fb2fc](https://github.com/nuxt/typescript/commit/48fb2fc))
+
+
+
+
+
+## [0.2.12](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.11...@nuxt/types@0.2.12) (2019-09-16)
+
+
+### Bug Fixes
+
+* **types:** allow boolean at render.compressor ([#115](https://github.com/nuxt/typescript/issues/115)) ([88fa988](https://github.com/nuxt/typescript/commit/88fa988))
+* **types:** export NuxtError to be able to annotate error prop ([#114](https://github.com/nuxt/typescript/issues/114)) ([7a0e539](https://github.com/nuxt/typescript/commit/7a0e539))
+
+
+
+
+
+## [0.2.11](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.10...@nuxt/types@0.2.11) (2019-09-06)
+
+
+### Bug Fixes
+
+* **types:** extendBuild instead of extendConfig ([e8961ff](https://github.com/nuxt/typescript/commit/e8961ff))
+
+
+
+
+
+## [0.2.10](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.9...@nuxt/types@0.2.10) (2019-09-06)
+
+
+### Bug Fixes
+
+* **types:** provide this.extendConfig type in modules ([91f3545](https://github.com/nuxt/typescript/commit/91f3545))
+* **types:** revert 79810725811b7ade83aa01decd3860e215785353 ([1d8b3c5](https://github.com/nuxt/typescript/commit/1d8b3c5))
+
+
+
+
+
+## [0.2.9](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.8...@nuxt/types@0.2.9) (2019-09-06)
+
+
+### Bug Fixes
+
+* **types:** optional build.additionalExtensions ([60c1dc4](https://github.com/nuxt/typescript/commit/60c1dc4))
+
+
+
+
+
+## [0.2.8](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.7...@nuxt/types@0.2.8) (2019-09-06)
+
+
+### Bug Fixes
+
+* **types:** missing types regarding extensions ([e9856b7](https://github.com/nuxt/typescript/commit/e9856b7))
+
+
+
+
+
+## [0.2.7](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.6...@nuxt/types@0.2.7) (2019-09-05)
+
+
+### Bug Fixes
+
+* **types:** configuration should be used as partial ([7981072](https://github.com/nuxt/typescript/commit/7981072))
+
+
+
+
+
+## [0.2.6](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.5...@nuxt/types@0.2.6) (2019-09-01)
+
+
+### Bug Fixes
+
+* **types:** add missing typings for NuxtApp interface ([#91](https://github.com/nuxt/typescript/issues/91)) ([7c57883](https://github.com/nuxt/typescript/commit/7c57883))
+* **types:** update types to match documentation examples ([#95](https://github.com/nuxt/typescript/issues/95)) ([b8064ce](https://github.com/nuxt/typescript/commit/b8064ce))
+
+
+
+
+
 ## [0.2.5](https://github.com/nuxt/typescript/compare/@nuxt/types@0.2.4...@nuxt/types@0.2.5) (2019-08-22)