+++ packages/typescript-build/CHANGELOG.md
@@ -3,6 +3,211 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.10...@nuxt/typescript-build@0.4.0) (2019-12-09)
+
+
+### Features
+
+* **deps:** upgrade typescript to 3.7 ([508fcef](https://github.com/nuxt/typescript/commit/508fcef3b3f0938edad9659849a82d89ac868c7c)), closes [#185](https://github.com/nuxt/typescript/issues/185)
+
+
+
+
+
+## [0.3.10](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.9...@nuxt/typescript-build@0.3.10) (2019-12-02)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.9](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.8...@nuxt/typescript-build@0.3.9) (2019-12-02)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.8](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.7...@nuxt/typescript-build@0.3.8) (2019-12-02)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.7](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.6...@nuxt/typescript-build@0.3.7) (2019-12-01)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.6](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.5...@nuxt/typescript-build@0.3.6) (2019-12-01)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.5](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.4...@nuxt/typescript-build@0.3.5) (2019-11-30)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.4](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.3...@nuxt/typescript-build@0.3.4) (2019-11-05)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.3](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.2...@nuxt/typescript-build@0.3.3) (2019-11-03)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.2](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.1...@nuxt/typescript-build@0.3.2) (2019-10-18)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.3.1](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.3.0...@nuxt/typescript-build@0.3.1) (2019-10-08)
+
+
+### Bug Fixes
+
+* **build:** bring back options.extensions statement ([52b3c3b](https://github.com/nuxt/typescript/commit/52b3c3b)), closes [#141](https://github.com/nuxt/typescript/issues/141)
+
+
+
+
+
+# [0.3.0](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.10...@nuxt/typescript-build@0.3.0) (2019-10-02)
+
+
+### Bug Fixes
+
+* **build:** fix feature [#133](https://github.com/nuxt/typescript/issues/133) after ts rewrite ([1e453e6](https://github.com/nuxt/typescript/commit/1e453e6))
+
+
+### Features
+
+* add custom blue color for nuxt cli ([#133](https://github.com/nuxt/typescript/issues/133)) ([3534073](https://github.com/nuxt/typescript/commit/3534073)), closes [#6397](https://github.com/nuxt/typescript/issues/6397)
+* use hooks ([#97](https://github.com/nuxt/typescript/issues/97)) ([f7b680a](https://github.com/nuxt/typescript/commit/f7b680a))
+
+
+
+
+
+## [0.2.10](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.9...@nuxt/typescript-build@0.2.10) (2019-09-23)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.9](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.8...@nuxt/typescript-build@0.2.9) (2019-09-17)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.8](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.7...@nuxt/typescript-build@0.2.8) (2019-09-16)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.7](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.6...@nuxt/typescript-build@0.2.7) (2019-09-16)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.6](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.5...@nuxt/typescript-build@0.2.6) (2019-09-06)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.5](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.4...@nuxt/typescript-build@0.2.5) (2019-09-06)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.4](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.3...@nuxt/typescript-build@0.2.4) (2019-09-06)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.3](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.2...@nuxt/typescript-build@0.2.3) (2019-09-06)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.2](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.1...@nuxt/typescript-build@0.2.2) (2019-09-05)
+
+**Note:** Version bump only for package @nuxt/typescript-build
+
+
+
+
+
+## [0.2.1](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.2.0...@nuxt/typescript-build@0.2.1) (2019-09-03)
+
+
+### Bug Fixes
+
+* pin typescript to 3.5.x ([d69ec27](https://github.com/nuxt/typescript/commit/d69ec27)), closes [#98](https://github.com/nuxt/typescript/issues/98)
+
+
+
+
+
+# [0.2.0](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.1.11...@nuxt/typescript-build@0.2.0) (2019-09-01)
+
+
+### Features
+
+* ts-loader options ([#93](https://github.com/nuxt/typescript/issues/93)) ([b80f6de](https://github.com/nuxt/typescript/commit/b80f6de))
+
+
+
+
+
 ## [0.1.11](https://github.com/nuxt/typescript/compare/@nuxt/typescript-build@0.1.10...@nuxt/typescript-build@0.1.11) (2019-08-22)
 
 **Note:** Version bump only for package @nuxt/typescript-build